* updated to new apigen and new tom backquote

+++ examples/mgs/BoulderDashApplet/JBoulderDashPanel.java
@@ -49,9 +49,9 @@ public class JBoulderDashPanel extends JPanel {
       // Warning between North and south,, east and west
     while(it.hasNext()) {
       Bead b = (Bead) it.next();
-      int x = b.getPos().getX().intValue();
-      int y = b.getPos().getY().intValue();
-      g.drawImage(img[b.getValue().intValue()], x*BoulderDashImagesData.IMAGE_SIZE, y*BoulderDashImagesData.IMAGE_SIZE, null);
+      int x = b.getPos().getX();
+      int y = b.getPos().getY();
+      g.drawImage(img[b.getValue()], x*BoulderDashImagesData.IMAGE_SIZE, y*BoulderDashImagesData.IMAGE_SIZE, null);
     }
     lastDrawnSpace = space;
   }
